+++ halo2_gadgets/CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `halo2_gadgets::sinsemilla::MessagePiece::from_subpieces`
+- `halo2_gadgets::utilities`:
+  - `FieldValue` trait.
+  - `RangeConstrained` newtype wrapper.
 
 ## [0.1.0-beta.3] - 2022-04-06
 ### Changed